+++ CHANGELOG.md
@@ -27,6 +27,10 @@ With this method, the version is YEAR.MONTH.RELEASE. To increment this version, 
 
 - Added `max_training_tokens` to specify the `max_length` parameter for the tokenizer when tokenizing the training set. Defaults to the prior setting of 512.
 
+## Changed
+
+- Logging: Switch from using logging module to loguru.
+
 ### Fixed
 
 - load_json_arg(): Fixed bug with json string as input.